refactor: add custom paths for user db functions
