+++ src/events/selectMenus/manageVotecount/addPlayers.ts
@@ -1,11 +1,11 @@
 import { type UserSelectMenuInteraction, type CacheType } from 'discord.js';
 import { UserSelectMenu } from '../../../structures/interactions/UserSelectMenu';
 import { InteractionError } from '../../../structures/interactions';
-import { getOrCreateUser, getPlayer, getVoteCounter } from '../../../util/database';
+import { getPlayer, getVoteCounter } from '../../../util/database';
 import { prisma } from '../../..';
 import { generateManagePlayersEmbed } from '../../buttons/manageVotecount/gotoPlayersMenu';
 import { manageVoteCountEmbeds } from '../../buttons/manageVotecount/goHome';
-
+import { getOrCreateUser } from '@models/users';
 export default class AddPlayersMenu extends UserSelectMenu {
 	static customId = 'manage-vc-players-add';
 	constructor() {
